Refactored project structure
Now MWScanner is a python module
Added a new Mixin for getting responses from URL
