+++ mwscanner/__init__.py
@@ -1,0 +1,3 @@
+# static global variable for all members of this module
+# it points to the Matricula Web website
+BASE_URL = 'https://matriculaweb.unb.br/'
